requirements.txt
hydra-core
captum
matplotlib
fiona
shapely
pyarrow 
pandas
openpyxl
geopandas 
earthengine-api
geemap
tqdm
pyyaml
plotly
scikit-learn
scikit-image
tensorboard
lightning
pip
iopath
segmentation-models-pytorch
h5py
einops
scipy
albumentations 
numpy
torchmetrics
py-xgboost
mlflow